Update src para compatibilidad con php src-4
