+++ composer.json
@@ -14,9 +14,11 @@
         "rmccue/requests": ">=1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.8.9"
+        "phpunit/phpunit": "^4.8.9"
     },
     "autoload": {
-        "classmap": ["epayco.php"]
+        "psr-4": {
+            "Epayco\\" : "src/"
+        }
     }
 }
