+++ src/Resources/Charge.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace Epayco\Resources;
+
+use Epayco\Resource;
+
+/**
+ * Charge payment methods
+ */
+class Charge extends Resource
+{
+    /**
+     * Create charge
+     * @param  object $options data charge
+     * @return object
+     */
+    public function create($options = null)
+    {
+        return $this->request(
+               "POST",
+               "/payment/v1/charge/create",
+               $api_key = $this->epayco->api_key,
+               $options,
+               $private_key = $this->epayco->private_key,
+               $test = $this->epayco->test,
+               $switch = false,
+               $lang = $this->epayco->lang
+        );
+    }
+
+    /**
+     * Return data transaction
+     * @param  String $uid id transaction
+     * @return object
+     */
+    public function transaction($uid = null)
+    {
+        return $this->request(
+                "GET",
+                "/restpagos/pse/transactioninfomation.json?transactionID=" . $uid . "&&public_key=" . $this->epayco->api_key,
+                $api_key = $this->epayco->api_key,
+                $uid,
+                $private_key = $this->epayco->private_key,
+                $test = $this->epayco->test,
+                $switch = true,
+                $lang = $this->epayco->lang
+        );
+    }
+}
