+++ src/Utils/key_lang.json
@@ -21,6 +21,5 @@
     "ip": "ip",
     "url_response": "url_respuesta",
     "url_confirmation": "url_confirmacion",
-    "method_confirmation": "metodoconfirmacion",
-    "end_date": "fechaexpiracion"
+    "method_confirmation": "metodoconfirmacion"
 }
